+++ cmd/parquet_receiver/parquet_receiver.go
@@ -1,4 +1,4 @@
-package examples
+package parquet_receiver
 
 import (
 	"errors"
