+++ cmd/text_receiver/text_receiver.go
@@ -1,4 +1,4 @@
-package examples
+package text_receiver
 
 import (
 	"bufio"
