--- dummy_client/go.mod
@@ -1,3 +1,0 @@
-module main
-
-go 1.22.1
