--- pgwatch3_rpc_server/README.md
@@ -1,5 +1,0 @@
-# RPC Receivers
-
-Primary type is the:
-
- - Write Request: Contains the file name, pgwatch3 Instance ID and MeasurementMessage
