--- pgwatch3_rpc_server/docker-compose.yml
@@ -1,5 +1,0 @@
-services:
-  rpc_sink:
-    build: .
-    ports:
-      - "5050:5050"
